+++ src/main/java/pl/asie/computronics/reference/Config.java
@@ -27,6 +27,7 @@ public class Config {
 	public static double FX_ENERGY_COST = 0.2;
 	public static double SOUND_ENERGY_COST = 1.0;
 	public static double SPOOFING_ENERGY_COST = 0.2;
+	public static double COLOR_CHANGE_COST = 0.2;
 	public static String CHATBOX_PREFIX = "ChatBox";
 	public static double LOCOMOTIVE_RELAY_RANGE = 128.0;
 	public static double LOCOMOTIVE_RELAY_BASE_POWER = 20.0;
@@ -44,6 +45,7 @@ public class Config {
 	public static boolean OC_CARD_SPOOF;
 	public static boolean OC_CARD_SOUND;
 	public static boolean OC_CARD_BOOM;
+	public static boolean OC_UPGRADE_COLORFUL;
 
 	public static boolean CC_OPEN_MULTI_PERIPHERAL = true;
 	public static boolean CC_ALL_MULTI_PERIPHERALS = true;
@@ -97,6 +99,7 @@ public class Config {
 			OC_CARD_SPOOF = config.get("enable.opencomputers", "spoofingCard", true).getBoolean(true);
 			OC_CARD_SOUND = config.get("enable.opencomputers", "soundCard", true).getBoolean(true);
 			OC_CARD_BOOM = config.get("enable.opencomputers", "boomCard", true).getBoolean(true);
+			OC_UPGRADE_COLORFUL = config.get("enable.opencomputers", "colorfulUpgrade", true).getBoolean(true);
 
 			// Particle Card
 			FX_ENERGY_COST = convertRFtoOC(
@@ -107,6 +110,9 @@ public class Config {
 			// Beep Card
 			SOUND_ENERGY_COST = convertRFtoOC(
 				config.getFloat("ocBeepCardCostPerSound", "power", 10.0f, 0.0f, 10000.0f, "How much energy a single beep will cost for 1 second"));
+			// Colorful Upgrade
+			COLOR_CHANGE_COST = convertRFtoOC(
+				config.getFloat("ocColorfulUpgradeColorChangeCost", "power", 2.0f, 0.0f, 10000.0f, "How much energy changing the color of the Colorful Upgrade will cost"));
 
 			if(Mods.isLoaded(Mods.Railcraft)) {
 				LOCOMOTIVE_RELAY_BASE_POWER = convertRFtoOC(